mcp-servers/mcp-cars/requirements.txt
amadeus>=11.0.0
python-dotenv>=1.0.0

